Fix writeframe generation for skinned meshes
